Move location tracking into its own file
